storage: add secured descriptor support
As per DSP0286, for the SPDM Storage Secured Messages contain
descriptors which allows a single Secured Message to transmit all elements
of a single command to an endpoint. Note that, within the  SPDM Storage
Secured Message data buffer, the fields from `Num Descriptors` through
`Command or Data Buffer`, inclusive, shall be treated as the Application Data
as described by DSP0277. Which means they must be encrypted at the
transport layer given the current design of libspdm.

Futhermore, in the context of libspdm, within the trasnport layer the the
storage specific (NVMe/SCSI/ATA) descriptors types are not particularly currently
relevent. However, to be compliant with DSP0286, we still need to
implement SPDM Storage Secured Message Descriptors supporting the `SPDM`
Descriptor type to encapsulate SPDM Messages in a secured session. The
`SPDM` type, describe an SPDM request or response message in secured
SPDM message.

This patch adds
`libspdm_strorage_secured_message_decode()/libspdm_strorage_secured_message_encode()`
to facilitate the above. The enocde function can be used to add add such a descriptor
to the SPDM message prior to it is encrypted, and similarly the decode function be
used after it is decrypted to parse the descriptor header.

Signed-off-by: Wilfred Mallawa <wilfred.mallawa@wdc.com>